+++ fractal_server/app/routes/api/v2/images.py
@@ -1,5 +1,3 @@
-from typing import Any
-
 from fastapi import APIRouter
 from fastapi import Depends
 from fastapi import HTTPException
@@ -26,6 +24,7 @@ from fractal_server.images.tools import aggregate_types
 from fractal_server.images.tools import find_image_by_zarr_url
 from fractal_server.images.tools import match_filter
 from fractal_server.types import AttributeFilters
+from fractal_server.types import ImageAttributeValue
 from fractal_server.types import TypeFilters
 
 router = APIRouter()
@@ -33,11 +32,19 @@ router = APIRouter()
 
 class ImagePage(PaginationResponse[SingleImage]):
 
-    attributes: dict[str, list[Any]]
+    attributes: dict[str, list[ImageAttributeValue]]
     types: list[str]
 
 
 class ImageQuery(BaseModel):
+    """
+    Query for a list of images.
+
+    Attributes:
+        type_filters:
+        attribute_filters:
+    """
+
     type_filters: TypeFilters = Field(default_factory=dict)
     attribute_filters: AttributeFilters = Field(default_factory=dict)