+++ fractal_server/app/schemas/v2/__init__.py
@@ -16,7 +16,6 @@ from .history import HistoryUnitRead  # noqa F401
 from .history import HistoryUnitStatus  # noqa F401
 from .history import HistoryUnitStatusWithUnset  # noqa F401
 from .history import ImageLogsRequest  # noqa F401
-from .history import SingleImageWithStatus  # noqa F401
 from .job import JobCreateV2  # noqa F401
 from .job import JobReadV2  # noqa F401
 from .job import JobStatusTypeV2  # noqa F401